ddl: remove db_legacy_test.go
