--- ddl/db_legacy_test.go
@@ -1,157 +1,0 @@
-// Copyright 2015 PingCAP, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package ddl_test
-
-import (
-	"context"
-	"testing"
-	"time"
-
-	. "github.com/pingcap/check"
-	"github.com/pingcap/errors"
-	"github.com/pingcap/tidb/ddl"
-	"github.com/pingcap/tidb/domain"
-	"github.com/pingcap/tidb/kv"
-	"github.com/pingcap/tidb/meta/autoid"
-	"github.com/pingcap/tidb/parser/model"
-	"github.com/pingcap/tidb/session"
-	"github.com/pingcap/tidb/sessionctx"
-	"github.com/pingcap/tidb/store/mockstore"
-	"github.com/pingcap/tidb/table"
-	"github.com/pingcap/tidb/util/testleak"
-	"github.com/stretchr/testify/require"
-	"github.com/tikv/client-go/v2/testutils"
-)
-
-func TestT(t *testing.T) {
-	CustomVerboseFlag = true
-	*CustomParallelSuiteFlag = true
-
-	testleak.BeforeTest()
-	TestingT(t)
-	testleak.AfterTestT(t)()
-}
-
-const (
-	// waitForCleanDataRound indicates how many times should we check data is cleaned or not.
-	waitForCleanDataRound = 150
-	// waitForCleanDataInterval is a min duration between 2 check for data clean.
-	waitForCleanDataInterval = time.Millisecond * 100
-)
-
-var _ = SerialSuites(&testSerialDBSuite{&testDBSuite{}})
-
-const defaultBatchSize = 1024
-const defaultReorgBatchSize = 256
-
-type testDBSuite struct {
-	cluster    testutils.Cluster
-	store      kv.Storage
-	dom        *domain.Domain
-	schemaName string
-	s          session.Session
-	lease      time.Duration
-	autoIDStep int64
-	ctx        sessionctx.Context
-}
-
-func setUpSuite(s *testDBSuite, c *C) {
-	var err error
-
-	s.lease = 600 * time.Millisecond
-	session.SetSchemaLease(s.lease)
-	session.DisableStats4Test()
-	s.schemaName = "test_db"
-	s.autoIDStep = autoid.GetStep()
-	ddl.SetWaitTimeWhenErrorOccurred(0)
-
-	s.store, err = mockstore.NewMockStore(
-		mockstore.WithClusterInspector(func(c testutils.Cluster) {
-			mockstore.BootstrapWithSingleStore(c)
-			s.cluster = c
-		}),
-	)
-	c.Assert(err, IsNil)
-
-	s.dom, err = session.BootstrapSession(s.store)
-	c.Assert(err, IsNil)
-	s.s, err = session.CreateSession4Test(s.store)
-	c.Assert(err, IsNil)
-	s.ctx = s.s.(sessionctx.Context)
-
-	_, err = s.s.Execute(context.Background(), "create database test_db")
-	c.Assert(err, IsNil)
-	_, err = s.s.Execute(context.Background(), "set @@global.tidb_max_delta_schema_count= 4096")
-	c.Assert(err, IsNil)
-}
-
-func tearDownSuite(s *testDBSuite, c *C) {
-	_, err := s.s.Execute(context.Background(), "drop database if exists test_db")
-	c.Assert(err, IsNil)
-	s.s.Close()
-	s.dom.Close()
-	err = s.store.Close()
-	c.Assert(err, IsNil)
-}
-
-func (s *testDBSuite) SetUpSuite(c *C) {
-	setUpSuite(s, c)
-}
-
-func (s *testDBSuite) TearDownSuite(c *C) {
-	tearDownSuite(s, c)
-}
-
-type testSerialDBSuite struct{ *testDBSuite }
-
-func testGetTableByName(c *C, ctx sessionctx.Context, db, table string) table.Table {
-	dom := domain.GetDomain(ctx)
-	// Make sure the table schema is the new schema.
-	err := dom.Reload()
-	c.Assert(err, IsNil)
-	tbl, err := dom.InfoSchema().TableByName(model.NewCIStr(db), model.NewCIStr(table))
-	c.Assert(err, IsNil)
-	return tbl
-}
-
-func backgroundExec(s kv.Storage, sql string, done chan error) {
-	se, err := session.CreateSession4Test(s)
-	if err != nil {
-		done <- errors.Trace(err)
-		return
-	}
-	defer se.Close()
-	_, err = se.Execute(context.Background(), "use test")
-	if err != nil {
-		done <- errors.Trace(err)
-		return
-	}
-	_, err = se.Execute(context.Background(), sql)
-	done <- errors.Trace(err)
-}
-
-func testGetIndexID(t *testing.T, ctx sessionctx.Context, dbName, tblName, idxName string) int64 {
-	is := domain.GetDomain(ctx).InfoSchema()
-	tt, err := is.TableByName(model.NewCIStr(dbName), model.NewCIStr(tblName))
-	require.NoError(t, err)
-
-	for _, idx := range tt.Indices() {
-		if idx.Meta().Name.L == idxName {
-			return idx.Meta().ID
-		}
-	}
-	require.FailNowf(t, "index %s not found(db: %s, tbl: %s)", idxName, dbName, tblName)
-	return -1
-}
