+++ testkit/external/util.go
@@ -15,6 +15,7 @@
 package external
 
 import (
+	"fmt"
 	"strings"
 	"testing"
 
@@ -25,16 +26,6 @@ import (
 	"github.com/pingcap/tidb/testkit"
 	"github.com/stretchr/testify/require"
 )
-
-// GetSchemaByName gets schema by name for test.
-func GetSchemaByName(t *testing.T, tk *testkit.TestKit, db string) *model.DBInfo {
-	dom := domain.GetDomain(tk.Session())
-	// Make sure the table schema is the new schema.
-	require.NoError(t, dom.Reload())
-	dbInfo, ok := dom.InfoSchema().SchemaByName(model.NewCIStr(db))
-	require.True(t, ok)
-	return dbInfo
-}
 
 // GetTableByName gets table by name for test.
 func GetTableByName(t *testing.T, tk *testkit.TestKit, db, table string) table.Table {
@@ -62,4 +53,20 @@ func GetModifyColumn(t *testing.T, tk *testkit.TestKit, db, tbl, colName string,
 		}
 	}
 	return nil
+}
+
+// GetIndexID is used to get the index ID from full qualified name.
+func GetIndexID(t *testing.T, tk *testkit.TestKit, dbName, tblName, idxName string) int64 {
+	is := domain.GetDomain(tk.Session()).InfoSchema()
+	tt, err := is.TableByName(model.NewCIStr(dbName), model.NewCIStr(tblName))
+	require.NoError(t, err)
+
+	for _, idx := range tt.Indices() {
+		if idx.Meta().Name.L == idxName {
+			return idx.Meta().ID
+		}
+	}
+
+	require.FailNow(t, fmt.Sprintf("index %s not found(db: %s, tbl: %s)", idxName, dbName, tblName))
+	return -1
 }
